modify: salePercentage type change

+++ src/main/java/com/munchymarket/MunchyMarket/dto/ProductDto.java
@@ -21,7 +21,7 @@ public class ProductDto {
     private int stock;
     private String mainImage;
     private Boolean isOnSale;
-    private BigDecimal salePercentage;
+    private int salePercentage;
     private Boolean isPurchaseStatus;
 
 }
